adjust gprc interceptor file struct
